+++ lab6/Triangle.hpp
@@ -15,13 +15,13 @@
 using namespace std;
 
 class Triangle {
-    vector<int> point1; //0-координата х, 1-у, 2-z
-    vector<int> point2;
-    vector<int> point3;
+    vector<float> point1; //0-координата х, 1-у, 2-z
+    vector<float> point2;
+    vector<float> point3;
     float color;
-    bool isIn2dTriangle(vector<vector<int>> trianglePoints, vector<int> findPoint);
+    bool isIn2dTriangle(vector<vector<float>> trianglePoints, vector<float> findPoint);
 public:
-    Triangle (int x1, int y1, int z1, int x2, int y2, int z2, int x3, int y3, int z3, float cosAlpha);
+    Triangle (float x1, float y1, float z1, float x2, float y2, float z2, float x3, float y3, float z3, float cosAlpha);
     float getColor();
     int getMax(int numCoor);
     int getMin(int numCoor);
